Final Commit for 1.2.5
For the last commit I added a way to end the game where the chips are counted and once you reach the 50k threshold the game ends/when your number of chips hits zero. After that I came across 2 different errors that was stopping the code from running well. First i didn't name all of my veriables correctly like hit and stand which was an easy fix that occured in a couple different spots. Secondly which has been the last hour and a half of my life the main loop for the game wasn't working where it was just close out right when I inputed my bet amount. After tinkering and searching through google and chatgpt the error was due to numinput returning as None and breaking the loop immediately and making my game window close along with it. I think i understnad what I changed but im going to look furhter into it tommorow so I can use this as a lesson for future referece. I also cleaned the UI up becasue the card slots were overlapping along with the text boxes which was an easy fix. Sorry this was so long.